requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.13
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
Babel==2.11.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-geoposition==0.3.0
django-google-maps==0.13.0
django-leaflet==0.28.3
django-messages==0.6.0
djangorestframework==3.14.0
docutils==0.17.1
et-xmlfile==1.1.0
exceptiongroup==1.1.0
frozenlist==1.3.3
future==0.18.2
geocoder==1.38.1
google==3.0.0
googlemaps==4.7.3
h11==0.14.0
html5lib==1.1
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.1
oauthlib==3.2.2
openpyxl==3.0.10
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.3.0
psycopg==3.1.7
psycopg2-binary==2.9.5
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
ratelim==0.1.6
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinx-rtd-theme==1.1.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
webencodings==0.5.1
wsproto==1.2.0
yarl==1.8.2
